Fix white screen on token expire
